% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_divergence.R
\name{amino_distance_matrix}
\alias{amino_distance_matrix}
\title{Get a pairwise 20x20 distance matrix for all pairs of amino acids}
\usage{
amino_distance_matrix(method = "grantham")
}
\arguments{
\item{method}{\code{"grantham"} for the Grantham 1974 matrix or \code{"uniform"} for a matrix with ones on the non-diagonal.}
}
\value{
A 20x20 symmetric matrix with positive numbers and zeros on the diagonal.
}
\description{
By default, we return the amino acid distance matrix by Grantham 1974 (doi:10.1126/science.185.4154.862).
}
\examples{
# By default, the Grantham 1974 matrix
amino_distance_matrix("grantham")

# All ones, and zeros on the diagonal
amino_distance_matrix("uniform")
}
