% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{amino_dosage}
\alias{amino_dosage}
\title{Amino acid dosage}
\usage{
amino_dosage(genotypes, aminos, drop_constants = TRUE, drop_duplicates = TRUE)
}
\arguments{
\item{genotypes}{Input character vector with one genotype for each individual.}

\item{aminos}{A one-hot encoded matrix with one row per allele and one
column per amino acid position.}

\item{drop_constants}{Filter out constant amino acid positions by default.}

\item{drop_duplicates}{Filter out duplicate amino acid positions by default.}
}
\value{
A data frame with one row for each input genotype.
}
\description{
For each genotype, return the the dosage for each amino acid at each
position.
}
\details{
Each genotype should be represented like \code{"HLA-A*01:01+HLA-A*01:01"}

By default, the returned data frame is filtered to exclude:
\itemize{
\item amino acid positions where all input genotypes have the same allele
\item amino acid positions that are identical to previous positions
}
}
