% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{dosage}
\alias{dosage}
\title{Convert a set of genotype names into a dosage matrix of each residue at each position}
\usage{
dosage(names, mat, drop_constants = TRUE, drop_duplicates = TRUE)
}
\arguments{
\item{names}{Input character vector with one genotype for each individual. All entries must be present in \code{rownames(mat)}.}

\item{mat}{A one-hot encoded matrix with one row per allele and one column for each residue (amino acid or nucleotide) at each position.}

\item{drop_constants}{Filter out constant amino acid positions by default.}

\item{drop_duplicates}{Filter out duplicate amino acid positions by default.}
}
\value{
A matrix with one row for each input genotype, and one column for each residue at each position.
}
\description{
For each genotype name, return the the dosage matrix for each residue (amino acid or nucleotide) at each position.
}
\details{
Each genotype should be represented like this \code{"HLA-A*01:01,HLA-A*01:01"}

By default, the returned matrix is filtered to exclude:
\itemize{
\item positions where all input genotypes have the same allele
\item positions that are identical to previous positions
}
}
\examples{
DRB1_file <- file.path(
  "https://github.com/ANHIG/IMGTHLA/raw",
  "5f2c562056f8ffa89aeea0631f2a52300ee0de17",
  "alignments/DRB1_prot.txt"
)
a <- read_alignments(DRB1_file)
genotypes <- c(
  "DRB1*12:02:02:03,DRB1*12:02:02:03,DRB1*14:54:02",
  "DRB1*04:174,DRB1*15:152",
  "DRB1*04:56:02,DRB1*15:01:48",
  "DRB1*14:172,DRB1*04:160",
  "DRB1*04:359,DRB1*04:284:02"
)
dosage <- dosage(genotypes, a$onehot)
dosage[,1:5]
}
