% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{get_onehot}
\alias{get_onehot}
\title{Make a one-hot encoded matrix from a dataframe of amino acid
sequences.}
\usage{
get_onehot(sequences, n_pre, drop_identical = FALSE, verbose = FALSE)
}
\arguments{
\item{n_pre}{The number of amino acid sequences before position 1.}

\item{drop_identical}{Drop positions where all alleles have the same amino acid. Default is false.}

\item{verbose}{Print messages along the way.}

\item{al}{A dataframe with columns allele, seq}
}
\description{
Make a one-hot encoded matrix from a dataframe of amino acid
sequences.
}
\keyword{internal}
