% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grantham}
\alias{grantham}
\title{Table 1 from Grantham 1974}
\format{
A data frame with 20 rows and 5 columns:
\describe{
\item{amino}{Anonymized Mechanical Turk Worker ID}
\item{c}{Composition c, the atomic weight ratio of noncarbon elements in end groups or rings to carbons in the side chain}
\item{p}{Polarity p from published data}
\item{v}{Volume v from published data}
}
}
\usage{
grantham
}
\description{
Grantham R. Amino Acid Difference Formula to Help Explain Protein Evolution.
Science. 1974;185: 862–864. doi:10.1126/science.185.4154.862
}
\examples{
grantham
}
\keyword{datasets}
