% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_genes}
\alias{hla_genes}
\title{Get HLA gene names from IMGTHLA}
\usage{
hla_genes(release = "latest", overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{release}{Default is "latest". Should be a release name like "3.51.0".}

\item{overwrite}{Overwrite the existing \code{genes.json} file with a new one from GitHub}

\item{verbose}{If TRUE, print messages along the way.}

\item{type}{Return \code{"{gene}_{type}"} (e.g. "DRB1_prot") instead of \code{"{gene}"} (e.g. "DRB1").}
}
\value{
A character vector of HLA gene names like "DRB1"
}
\description{
Retrieve the list of txt files in \href{https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments}{\code{github.com/ANHIG/IMGTHLA/alignments}} and return a list of gene names derived from the file names.
}
\examples{
\donttest{
hla_genes() 
}
}
\seealso{
\code{\link[=hla_releases]{hla_releases()}} to get a complete list of all release names.
}
