% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_releases}
\alias{hla_releases}
\title{Get IMGTHLA releases from GitHub}
\usage{
hla_releases(overwrite = FALSE)
}
\value{
A character vector of release names like "3.51.0"
}
\description{
Get IMGTHLA releases from GitHub
}
