% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud-package.R
\docType{package}
\name{hlabud-package}
\alias{hlabud}
\alias{hlabud-package}
\title{hlabud: Amino Acid Association Methods for the Human Leukocyte Antigen (HLA) Genes}
\description{
Methods for analysis of HLA genes at the amino acid level.
}
\author{
\strong{Maintainer}: Kamil Slowikowski \email{kslowikowski@gmail.com} (\href{https://orcid.org/0000-0002-2843-6370}{ORCID})

}
\keyword{internal}
