% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{one_to_three}
\alias{one_to_three}
\title{Convert one letter amino acid codes to three letter amino acid codes}
\usage{
one_to_three(aminos)
}
\description{
Convert one letter amino acid codes to three letter amino acid codes
}
\keyword{internal}
