% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{read_alignment}
\alias{read_alignment}
\title{Read an alignment file \verb{*_prot.txt} file from IMGTHLA.}
\usage{
read_alignment(my_file)
}
\arguments{
\item{my_file}{File name for a txt file from IMGTHLA like "DQB1_prot.txt"}
}
\value{
A list with a dataframe and a matrix. The dataframe has two columns:
\itemize{
\item allele: the name of the allele, e.g., \code{DQB*01:01}
\item seq: the amino acid sequence
The matrix has a one-hot encoding of the variants among the alleles, with
one row for each allele and one column for each amino acid at each position.
}
}
\description{
The prot file has the amino acid sequence for each HLA allele.
}
\examples{
my_file <- file.path(
  "https://github.com/ANHIG/IMGTHLA/raw",
  "5f2c562056f8ffa89aeea0631f2a52300ee0de17",
  "alignments/DRB1_prot.txt"
)
a <- read_alignment(my_file)
head(a$sequences)
a$alleles[1:5,1:5]
a$onehot[1:5,1:5]
}
