% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{read_prot}
\alias{read_prot}
\title{Read a \verb{*_prot.txt} file from IMGTHLA.}
\usage{
read_prot(prot_file)
}
\value{
A list with a dataframe and a matrix. The dataframe has two columns:
\itemize{
\item allele: the name of the allele, e.g., \code{DQB*01:01}
\item seq: the amino acid sequence
The matrix has a one-hot encoding of the variants among the alleles, with
one row for each allele and one column for each amino acid at each position.
}
}
\description{
Read a \verb{*_prot.txt} file from IMGTHLA.
}
