% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{amino_dosage}
\alias{amino_dosage}
\title{Amino acid dosage}
\usage{
amino_dosage(genotypes, aminos, drop_constants = TRUE, drop_duplicates = TRUE)
}
\arguments{
\item{genotypes}{Input character vector with one genotype for each individual.}

\item{aminos}{A one-hot encoded matrix with one row per allele and one
column per amino acid position.}

\item{drop_constants}{Filter out constant amino acid positions by default.}

\item{drop_duplicates}{Filter out duplicate amino acid positions by default.}
}
\value{
A data frame with one row for each input genotype.
}
\description{
For each genotype, return the the dosage for each amino acid at each
position.
}
\details{
Each genotype should be represented like \code{"HLA-A*01:01+HLA-A*01:01"}

By default, the returned data frame is filtered to exclude:
\itemize{
\item amino acid positions where all input genotypes have the same allele
\item amino acid positions that are identical to previous positions
}
}
\examples{
DRB1_file <- file.path(
  "https://github.com/ANHIG/IMGTHLA/raw",
  "5f2c562056f8ffa89aeea0631f2a52300ee0de17",
  "alignments/DRB1_prot.txt"
)
a <- read_prot(DRB1_file)
genotypes <- c(
  "DRB1*12:02:02:03,DRB1*12:02:02:03,DRB1*14:54:02",
  "DRB1*04:174,DRB1*15:152",
  "DRB1*04:56:02,DRB1*15:01:48",
  "DRB1*14:172,DRB1*04:160",
  "DRB1*04:359,DRB1*04:284:02"
)
dosage <- amino_dosage(genotypes, a$aminos)
dosage[,1:5]
}
