% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{get_hlabud_dir}
\alias{get_hlabud_dir}
\title{Get the name of the folder for caching downloaded IMGTHLA files}
\usage{
get_hlabud_dir()
}
\value{
The name of the folder.
}
\description{
Get the folder name from \code{getOption("hlabud_dir")} or else use the
\href{https://github.com/r-lib/rappdirs}{rappdirs} package to choose an
appropriate folder for your operating system. The folder will be created
automatically if it does not exist. And the \code{hlabud_dir} option will bet set
to that folder.
}
\details{
Linux:

\if{html}{\out{<div class="sourceCode">}}\preformatted{~/.local/share/hlabud
}\if{html}{\out{</div>}}

Mac:

\if{html}{\out{<div class="sourceCode">}}\preformatted{~/Library/Application Support/hlabud
}\if{html}{\out{</div>}}

Windows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{C:\\Documents and Settings\\<User>\\Application Data\\slowkow\\hlabud
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
hlabud_dir <- get_hlabud_dir()
}
}
