% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{get_onehot}
\alias{get_onehot}
\title{Make a one-hot encoded matrix from a dataframe of amino acid
sequences.}
\usage{
get_onehot(al, n_pre)
}
\arguments{
\item{al}{A dataframe with columns allele, seq}

\item{n_pre}{The number of amino acid sequences before position 1.}
}
\description{
Make a one-hot encoded matrix from a dataframe of amino acid
sequences.
}
\keyword{internal}
