% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_alignments}
\alias{hla_alignments}
\title{Get aligned sequences in a dataframe}
\usage{
hla_alignments(release = NULL, gene = "DRB", type = "prot")
}
\value{
A dataframe.
}
\description{
Get aligned sequences in a dataframe
}
