% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_alignments}
\alias{hla_alignments}
\title{Get aligned sequences in a dataframe}
\usage{
hla_alignments(gene = "DRB1", type = "prot", release = NULL, quiet = TRUE)
}
\arguments{
\item{gene}{The name of a gene like "DRB1"}

\item{type}{The type of sequence, one of "prot", "nuc", "gen"}

\item{release}{The name of a release like "3.51.0"}

\item{quiet}{If FALSE, print messages along the way.}
}
\value{
A dataframe.
}
\description{
Get aligned sequences in a dataframe
}
