% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_alignments}
\alias{hla_alignments}
\title{Get aligned sequences in a dataframe}
\usage{
hla_alignments(
  gene = "DRB1",
  type = "prot",
  release = "latest",
  verbose = FALSE
)
}
\arguments{
\item{gene}{The name of a gene like "DRB1"}

\item{type}{The type of sequence, one of "prot", "nuc", "gen"}

\item{release}{Default is "latest". Should be a release name like "3.51.0".}

\item{verbose}{If TRUE, print messages along the way.}
}
\value{
A dataframe.
}
\description{
Here are the conventions used for alignments (\href{https://www.ebi.ac.uk/ipd/imgt/hla/alignment/help/}{EBI IMGT-HLA help page}):
\itemize{
\item The entry for each allele is displayed in respect to the reference sequences.
\item Where identity to the reference sequence is present the base will be displayed as a hyphen (-).
\item Non-identity to the reference sequence is shown by displaying the appropriate base at that position.
\item Where an insertion or deletion has occurred this will be represented by a period (.).
\item If the sequence is unknown at any point in the alignment, this will be represented by an asterisk (*).
\item In protein alignments for null alleles, the 'Stop' codons will be represented by a hash (X).
\item In protein alignments, sequence following the termination codon, will not be marked and will appear blank.
\item These conventions are used for both nucleotide and protein alignments.
}
}
\examples{
\donttest{
a <- hla_alignments("DRB1")
head(a$sequences)
a$alleles[1:6,1:6]
a$onehot[1:6,1:6]
}
}
\seealso{
\code{\link[=hla_releases]{hla_releases()}} to get a complete list of all release names.
}
