% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_alleles}
\alias{hla_alleles}
\title{Get a table of allele names for a particular IMGTHLA release}
\usage{
hla_alleles(release = "latest", overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{release}{Default is "latest". Should be a release name like "3.51.0".}

\item{overwrite}{Overwrite the existing \code{alleles.json} file and \verb{Allelelist.\{version\}.txt} file}

\item{verbose}{If TRUE, print messages along the way.}
}
\value{
A data frame with HLA allele ids and names
}
\description{
Download a list of all allele names for all HLA genes for a particular
IMGTHLA release.
}
\examples{
\donttest{
head(hla_alleles())
}
}
\seealso{
\code{\link[=hla_releases]{hla_releases()}} to get a complete list of all release names.
}
