% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_genes}
\alias{hla_genes}
\title{Get IMGTHLA gene names}
\usage{
hla_genes(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrite the existing \code{genes.json} file with a new one from GitHub}
}
\value{
A character vector of HLA gene names like "DRB1"
}
\description{
Get IMGTHLA gene names
}
\examples{
\donttest{
hla_genes() 
}
}
