% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_genes}
\alias{hla_genes}
\title{Get IMGTHLA gene names}
\usage{
hla_genes(release = "latest", overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{release}{Default is "latest". Should be a release name like "3.51.0".}

\item{overwrite}{Overwrite the existing \code{genes.json} file with a new one from GitHub}

\item{verbose}{If TRUE, print messages along the way.}
}
\value{
A character vector of HLA gene names like "DRB1"
}
\description{
Retrieve the contents of \href{https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments}{\code{github.com/ANHIG/IMGTHLA/alignments}} and return a list of gene names derived from the txt files in that folder.
}
\examples{
\donttest{
hla_genes() 
}
}
\seealso{
\code{\link[=hla_releases]{hla_releases()}} to get a complete list of all release names.
}
