% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_releases}
\alias{hla_releases}
\title{Get IMGTHLA releases from GitHub}
\usage{
hla_releases(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrite the existing tags.json file in \code{file.path(getOption("hlabud_dir"), "tags.json")}with a new one from the \href{https://github.com/ANHIG/IMGTHLA}{IMGTHLA} GitHub repo}
}
\value{
A character vector of release names like "3.51.0"
}
\description{
Retrieve the tags from the ANHIG/IMGTHLA repo and get the associated release
names.
}
\examples{
\donttest{
hla_releases() 
}
}
