% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{hla_releases}
\alias{hla_releases}
\title{Get the names of releases from IMGTHLA}
\usage{
hla_releases(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrite the existing tags.json file in \code{getOption("hlabud_dir")}}
}
\value{
A character vector of release names like "3.51.0"
}
\description{
Get tags from \href{https://github.com/ANHIG/IMGTHLA}{github.com/ANHIG/IMGTHLA}, save them to a file called \code{tags.json} in \code{getOption("hlabud_dir")},  and return the release names from that file.
}
\details{
The \code{tags.json} file will be automatically overwritten if it is older than 24 hours.
}
\examples{
\donttest{
hla_releases() 
}
}
