% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud-package.R
\docType{package}
\name{hlabud-package}
\alias{hlabud}
\alias{hlabud-package}
\title{hlabud: Methods for Access and Analysis of the Human Leukocyte Antigen (HLA) Gene Sequence Alignments from IMGTHLA}
\description{
Fetch sequence alignment data from the IMGTHLA database by Robinson et al (2020) \doi{10.1093/nar/gkz950}, automatically convert the sequence alignments into convenient R matrices ready for downstream analysis. The vignette shows examples of using the one-hot encoding of the data with logistic regression and dimensionality reduction. Data is downloaded lazily, as-needed, and cached in a user-configurable folder.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/slowkow/hlabud}
  \item \url{https://slowkow.github.io/hlabud}
  \item Report bugs at \url{https://github.com/slowkow/hlabud/issues}
}

}
\author{
\strong{Maintainer}: Kamil Slowikowski \email{kslowikowski@gmail.com} (\href{https://orcid.org/0000-0002-2843-6370}{ORCID})

}
\keyword{internal}
