% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud-package.R
\docType{package}
\name{hlabud-package}
\alias{hlabud}
\alias{hlabud-package}
\title{hlabud: Methods for Access and Analysis of the Human Leukocyte Antigen (HLA) Gene Sequence Alignments from IMGTHLA}
\description{
hlabud provides methods for retrieving sequence alignment data from the IMGTHLA database by Robinson et al (2020) \doi{10.1093/nar/gkz950}, and for converting the sequence alignments into R matrices ready for downstream analysis. Data is downloaded lazily, as-needed, and cached in a user-configurable folder. The vignette shows examples of how to analyze the data with logistic regression and dimensionality reduction.
}
\author{
\strong{Maintainer}: Kamil Slowikowski \email{kslowikowski@gmail.com} (\href{https://orcid.org/0000-0002-2843-6370}{ORCID})

}
\keyword{internal}
