% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud-package.R
\docType{package}
\name{hlabud-package}
\alias{hlabud}
\alias{hlabud-package}
\title{hlabud: Methods for Accessing the Human Leukocyte Antigen (HLA) Gene Alignments from IMGTHLA}
\description{
hlabud provides methods for retrieving sequence alignment data from the IMGTHLA database, and for converting the sequence alignments into R matrices ready for downstream analysis. The vignette shows examples of how to use the data with logistic regression and dimensionality reduction.
}
\author{
\strong{Maintainer}: Kamil Slowikowski \email{kslowikowski@gmail.com} (\href{https://orcid.org/0000-0002-2843-6370}{ORCID})

}
\keyword{internal}
