% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{install_hla}
\alias{install_hla}
\title{Download and unpack a tarball release of IMGTHLA from GitHub}
\usage{
install_hla(release = "latest", quiet = FALSE)
}
\description{
Fetch a list of releases from Github. By default, install the latest
release to the folder appropriate for your operating system:
}
\details{
Linux:
~/.local/share/hlabud
Mac:
~/Library/Application Support/hlabud
Windows:
C:\Documents and Settings\\if{html}{\out{<User>}}\Application Data\slowkow\hlabud

A typical release is about 120 MB in size, so the download can take a few
minutes.

We can get or set the folder with the \code{hlabud_dir} option:

my_dir <- getOption("hlabud_dir")
options(hlabud_dir = my_dir)

The release tarball from Github is unpacked into the folder.

Other functions in the hlabud package will use the unpacked data, or else
they will automatically download the minimum necessary files.
}
