% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{install_hla}
\alias{install_hla}
\title{Download and unpack a tarball release of IMGTHLA from GitHub}
\usage{
install_hla(release = "latest", overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{release}{Default is "latest". Should be a release name like "3.51.0".}

\item{overwrite}{If TRUE, overwrite existing files in the release folder.}

\item{verbose}{If TRUE, print messages along the way.}
}
\description{
The release tarball from Github is unpacked into the
\code{getOption("hlabud_dir")} folder.
}
\details{
Note that the latest releases are more than 100 MB in size, so the download
might take a while on slow connections.
}
\examples{
\dontrun{
install_hla()
install_hla("3.51.0")
install_hla("3.51.0", verbose = TRUE)
# Change the install directory
options(hlabud_dir = "path/to/my/dir")
install_hla()
}
}
