% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlabud.R
\name{read_prot}
\alias{read_prot}
\title{Read a \verb{*_prot.txt} file from IMGTHLA.}
\usage{
read_prot(prot_file)
}
\arguments{
\item{prot_file}{File name for a txt file from IMGTHLA like "DQB1_prot.txt"}
}
\value{
A list with a dataframe and a matrix. The dataframe has two columns:
\itemize{
\item allele: the name of the allele, e.g., \code{DQB*01:01}
\item seq: the amino acid sequence
The matrix has a one-hot encoding of the variants among the alleles, with
one row for each allele and one column for each amino acid at each position.
}
}
\description{
Read a \verb{*_prot.txt} file from IMGTHLA.
}
\examples{
DRB1_file <- file.path(
  "https://github.com/ANHIG/IMGTHLA/raw",
  "5f2c562056f8ffa89aeea0631f2a52300ee0de17",
  "alignments/DRB1_prot.txt"
)
a <- read_prot(DRB1_file)
head(a$sequences)
head(a$aminos)
head(a$onehot)
}
